+++ cueforge/__init__.py
@@ -2,4 +2,4 @@
 # Copyright (C) 2026 Lesani. See LICENSE for details.
 """CueForge - audio cue playback system for live theater."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"